Change nginx_version scheme
This changeset changes the version scheme for the Nginx installation
from package version numbers to PPA type.

Whenever a new package is released for Nginx in the PPA used by this
role, it doesn't get added to the repository, instead it replaces the
existing package. This becomes problematic if you are attempting to pin
package versions.

The best alternative I could come up with was to pin just the repository
type (development or stable). This way, if package version number
`1.6.2-2+trusty0` gets replaced with `1.6.2-4+trusty0`, the role doesn't
break.

In addition, this package creates a stub site in `/var/www/html`. If
`delete_default_site` is truthy, and the default site configuration was
just removed, remove the web root stub as well.